requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
reportlab
qrcode[pil]
python-dotenv
